pyproject.toml
[tool.poetry]
name = "reweave"
version = "0.2.0"
description = "AI-powered content creation toolkit for automated video, image, and text generation"
authors = ["Chandresh Kumar <chandresh.code@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
google-genai = "^1.0.0"
gtts = "^2.5.0"
moviepy = "^2.0"
pydantic = "^2.6.3"
python-dotenv = "^1.1.0"
youtube-transcript-api = "^1.0.0"
yt-dlp = "^2025.6.9"
pytest = "^8.4.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
